initial vulkan port
 - PlotScene: base vulkan plot renderer using QSGRenderNode implementation
 - PlotSceneItem: plot scene component (grid, line, etc...)

+++ QMcuPlot/tests/TestCounter.hpp
@@ -12,7 +12,7 @@ class TestCounter : public AbstractPlotDataProvider
   Q_OBJECT
   QML_ELEMENT
 
-  using base_type = int32_t;
+  using base_type = float;
 
   Q_PROPERTY(int initial MEMBER initial_)
   Q_PROPERTY(int low MEMBER low_)
